+++ OUPharmacyManagementApp/mainApp/viewsets/__init__.py
@@ -1,0 +1,2 @@
+from .bill import  BillViewSet
+from .doctor_schedule import DoctorScheduleViewSet
